New Methods in MovieCollection.java
editAMovie: a method that allows the user to edit a movie of the users choice.

editCharacterInAMovie: a method that allows the user to edit a character in a movie of the users choice.

displaySpecificCharacter: a method that displays the characters in a specific movie entered by the user.

Setters added to MonsterMovie.java

New cases added in Main.java

New javadocs along with edited javadocs